--- a/Misc/NEWS.d/next/Library/2024-08-06-10-36-55.gh-issue-118761.q_x_1A.rst
+++ b/Misc/NEWS.d/next/Library/2024-08-06-10-36-55.gh-issue-118761.q_x_1A.rst
@@ -1,0 +1,2 @@
+Improve import time of :mod:`pprint` by around seven times. Patch by Hugo
+van Kemenade.