<!--
Thanks for your contribution!
Please read this comment in its entirety. It's quite important.

# Pull Request title

It should be in the following format:

```
gh-NNNNN: Summary of the changes made
```

Where: gh-NNNNN refers to the GitHub issue number.

Most PRs will require an issue number. Trivial changes, like fixing a typo, do not need an issue.

# Backport Pull Request title

If this is a backport PR (PR made against branches other than `main`),
please ensure that the PR title is in the following format:

```
[X.Y] <title from the original PR> (GH-NNNN)
```

Where: [X.Y] is the branch name, e.g. [3.6].

GH-NNNN refers to the PR number from `main`.

-->

Improves import time by up to seven times. Measured with a PGO and LTO non-debug build on macOS.

* The slowest import in `pprint` is `dataclasses` (taking 0.0012s / 55.2% of import time), and used in only two functions. 

* After deferring `dataclasses`, the next is `re` (0.005s / 30.1% of the new import time), only used in one function.

* After deferring `re`, the next are `collections` (0.002s / 14.8%) and `types` (0.000s / 2.4%). I left these, as their times are low and they're less straightforward to defer.

## `-X importtime`

Running: `pip install tuna && ./python.exe -X importtime -c 'import pprint' 2> import.log && tuna import.log`

### total import time: 0.022s -> 0.011s = x2 as fast

<table>
<tr>
<th><code>main</code>
<th>PR
<tr>
<td><img width="1624" alt="image" src="https://github.com/user-attachments/assets/4c37ccd5-21b7-43a4-bb2e-6c2bfedad0f3">
<td><img width="1624" alt="image" src="https://github.com/user-attachments/assets/f570333a-7cb5-47c6-9d23-e49c551bbc23">
</table>

### `pprint` import time: 0.014s -> 0.002s = x7 as fast

<table>
<tr>
<th><code>main</code>
<th>PR
<tr>
<td><img width="1624" alt="image" src="https://github.com/user-attachments/assets/02e7983a-8b74-4766-8cfa-2a683582676b">
<td><img width="1624" alt="image" src="https://github.com/user-attachments/assets/7ec02191-025d-47d5-b53b-19be10ee2bba">
</table>



## hyperfine: 19.2 ms -> 10.4 ms = x1.8 as fast

### main

```console
❯ hyperfine -w 8 "./python.exe -c 'import pprint'"
Benchmark 1: ./python.exe -c 'import pprint'
  Time (mean ± σ):      19.2 ms ±   0.7 ms    [User: 15.8 ms, System: 2.8 ms]
  Range (min … max):    18.1 ms …  21.3 ms    135 runs
```

### PR

```console
❯ hyperfine -w 8 "./python.exe -c 'import pprint'"
Benchmark 1: ./python.exe -c 'import pprint'
  Time (mean ± σ):      10.4 ms ±   0.4 ms    [User: 8.1 ms, System: 1.8 ms]
  Range (min … max):     9.8 ms …  12.4 ms    212 runs
```

Also add some PEP 8 whitespace to remove some distracting IDE warnings.


<!-- gh-issue-number: gh-118761 -->
* Issue: gh-118761
<!-- /gh-issue-number -->