@hugovk 2024-08-06
By the way, what's the reason for importing with underscores?

```python
import collections as _collections
import sys as _sys
import types as _types
from io import StringIO as _StringIO
```